Unstack Kore.Parser hierarchy
